Fix offline map store circular dependencies
